Move maybes into their own file
